Fix aarch64-apple-darwin build of the LZ4 frame bound
`std::min` deduces a single `_Tp` from both arguments, so they must have the same
type rather than merely be convertible. `max_block_size` was `size_t` while
`decoded.value_or(...)` is `UInt64`. On Linux x86_64 and aarch64 both spell
`unsigned long`, so deduction succeeds and all 12 Linux builds compile; on
aarch64-apple-darwin `UInt64` is `unsigned long long` and `size_t` is
`unsigned long`, so deduction fails and the build stops with -Werror.

Declare `max_block_size` as `UInt64` rather than annotating the call: `bound` is
already `UInt64` and accumulates it, and the `LZ4F_max*` cases are small
constants, so this makes the whole function consistent instead of widening at one
call site. Both types are 64-bit and unsigned on LP64, and `block_size` is at most
4 MiB so the multiplication cannot wrap, leaving Linux behaviour unchanged.

The public CI did not catch this because no darwin build ran on this branch.
Verified by compiling the function under darwin's type model, where the reverted
declaration reproduces the reported error and the fix compiles; the whole
translation unit is clean under that model.

+++ src/Processors/Formats/Impl/ArrowIPC/BufferCompression.cpp
@@ -96,7 +96,7 @@ UInt64 lz4BlockOutputBound(const LZ4F_frameInfo_t & info, const char * src, size
 {
     /// `LZ4F_getBlockSize` maps these, but is declared only in the static-linking section of
     /// `lz4frame.h`, which does not compile here. The parsed header rejects any other value.
-    size_t max_block_size = 0;
+    UInt64 max_block_size = 0;
     switch (info.blockSizeID)
     {
         case LZ4F_default: [[fallthrough]];
